# openapi.yaml
openapi: 3.0.3
info:
  title: Karton API
  description: >
    API za aplikaciju "Karton" — vođenje auto servisa. Backend je jedini autoritet
    za poslovna pravila (specifikacija-finalna.md §7). Nazivi na engleskom (odluka 27),
    poruke grešaka na srpskom. Sva vremena u UTC; poslovna logika Europe/Belgrade.

    CSRF ugovor: svaki odgovor postavlja/osvežava XSRF-TOKEN cookie (čitljiv iz JS);
    svaki non-GET zahtev MORA poslati header X-CSRF-Token sa tom vrednošću —
    neslaganje ili izostanak → 403 CSRF_FAILED. Važi i za /auth/login (token se
    dobija prethodnim GET pozivom, npr. /auth/me).

    Optimističko zaključavanje: verzionisani agregati su work_order, document,
    appointment i settings. Sve mutacione akcije nad njima nose "version" u telu;
    zastarela verzija → 409 VERSION_CONFLICT. Akcije bez version polja (copy,
    DELETE termina, izdavanje predračuna) backend izvršava u transakciji uz
    row-lock i atomsku proveru aktuelnog stanja.
  version: 1.0.0
servers:
  - url: /api/v1
security:
  - cookieAuth: []
...
paths:
...
  /reports/revenue:
    get:
      tags: [reports]
      summary: Prihod — samo plaćeni računi po datumu plaćanja (BR-31); grafik po mesecima
      parameters:
        - { name: from, in: query, schema: { type: string, format: date } }
        - { name: to, in: query, schema: { type: string, format: date } }
        - { name: format, in: query, schema: { type: string, enum: [json, xlsx] } }
      responses: { '200': { description: OK } }
  /reports/work-orders:
    get:
      tags: [reports]
      summary: Pretraga naloga (klijent, vozilo, tablica, marka, model; delovi kao tagovi)
      parameters:
        - $ref: '#/components/parameters/Query'
        - $ref: '#/components/parameters/Page'
        - { name: format, in: query, schema: { type: string, enum: [json, xlsx] } }
      responses: { '200': { description: OK } }
  /reports/mechanics/{id}:
    get:
      tags: [reports]
      summary: Po majstoru — nalozi, vrednost rada, period
      parameters:
        - $ref: '#/components/parameters/Id'
        - { name: from, in: query, schema: { type: string, format: date } }
        - { name: to, in: query, schema: { type: string, format: date } }
      responses: { '200': { description: OK } }
  /reports/vehicle-types:
    get:
      tags: [reports]
      summary: Po tipu vozila (marka, model, godina, gorivo)
      parameters:
        - { name: make, in: query, schema: { type: string } }
        - { name: model, in: query, schema: { type: string } }
        - { name: year, in: query, schema: { type: integer } }
        - { name: fuel, in: query, schema: { type: string } }
      responses: { '200': { description: OK } }

  # ---------- Export ----------
...
  /settings:
    get:
      tags: [settings]
      responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Settings' } } } }, '403': { $ref: '#/components/responses/Forbidden' } }
    patch:
      tags: [settings]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/Settings' } } } }
      responses: { '200': { description: OK }, '403': { $ref: '#/components/responses/Forbidden' }, '409': { $ref: '#/components/responses/Conflict' } }
...
  /settings/logo:
    put:
      tags: [settings]
      summary: 'ADMIN: postavi logotip (data URI, do 400 KB; PNG/JPEG/SVG/WebP)'
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [dataUrl], properties: { dataUrl: { type: string, example: 'data:image/png;base64,iVBORw0KGgo...' } } } } }
      responses: { '200': { description: Podešavanja sa novim logotipom }, '400': { $ref: '#/components/responses/Validation' }, '403': { description: Nije administrator } }
    delete:
      tags: [settings]
      summary: 'ADMIN: ukloni logotip'
      responses: { '200': { description: Podešavanja bez logotipa } }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        '200': { description: Sačuvan, content: { application/json: { schema: { type: object, properties: { logo: { type: string } } } } } }
        '403': { $ref: '#/components/responses/Forbidden' }
  /users:
    get:
      tags: [settings]
      responses: { '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/User' } } } } }, '403': { $ref: '#/components/responses/Forbidden' } }
    post:
      tags: [settings]
      summary: 'ADMIN: novi korisnik (nema samoregistracije)'
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/UserInput' } } } }
      responses: { '201': { description: Kreiran }, '409': { $ref: '#/components/responses/Duplicate' } }
  /users/{id}:
    parameters: [ { $ref: '#/components/parameters/Id' } ]
    patch:
      tags: [settings]
      summary: 'ADMIN: izmena, rola, status, reset lozinke'
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/UserInput' } } } }
      responses: { '200': { description: OK } }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: karton_session
  parameters:
    Id: { name: id, in: path, required: true, schema: { type: integer } }
...
    Page: { name: page, in: query, schema: { type: integer, default: 1 } }
...
    Query: { name: q, in: query, schema: { type: string }, description: Case-insensitive, ignoriše razmake }
  responses:
...
    Forbidden:
      description: Rola nema pravo (FORBIDDEN)
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Conflict:
      description: VERSION_CONFLICT — optimističko zaključavanje
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Duplicate:
      description: DUPLICATE_VIN / DUPLICATE_TAX_ID / DUPLICATE_EMAIL — nosi existingId
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string, example: PROFORMA_NOT_VALID }
        message: { type: string, example: 'Predračun je istekao i ne može se pretvoriti u račun.' }
        fields: { type: object, additionalProperties: { type: string } }
        warnings: { type: array, items: { type: string }, example: [MECHANIC_BUSY] }
        existingId: { type: integer, description: Kod duplikata — ID postojećeg entiteta }
...
    User:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string }
        role: { type: string, enum: [admin, user] }
        status: { type: string, enum: [active, disabled] }
    UserInput:
      type: object
      properties:
        name: { type: string }
        email: { type: string }
        password: { type: string }
        role: { type: string, enum: [admin, user] }
        status: { type: string, enum: [active, disabled] }
...
    Settings:
      type: object
      properties:
        shopName: { type: string }
        address: { type: string }
        taxId: { type: string }
        phone: { type: string }
        logo: { type: string }
        smtpHost: { type: string }
        smtpPort: { type: integer }
        smtpUsername: { type: string }
        smtpPassword: { type: string, writeOnly: true }
        senderEmail: { type: string }
        workHoursFrom: { type: string, example: '08:00' }
        workHoursTo: { type: string, example: '18:00' }
        defaultValidityDays: { type: integer, default: 15 }
        reminderSendTime: { type: string, example: '09:00' }
        pageSize: { type: integer, default: 20 }
        version: { type: integer }